fix: image load jumping
The website appears to jump as the user scrolls down and up because the
image pop in and out as they leave the view (and the replacement doesn't
retain the size). This PR removes the inview logic as a temporary
solution to improving the usability. We'll want to consider rewriting
the logic (per #397) to improve the load times.

Closes #404